Task_0

+++ 0x1E-search_algorithms/0-linear.c
@@ -14,9 +14,9 @@
 
 int linear_search(int *array, size_t size, int value)
 {
-	unsigned int index;
+	int index;
 
-	for (index = 0; array && index < size; index++)
+	for (index = 0; array && index < (int)size; index++)
 	{
 		printf("Value checked array[%d] = [%d]\n", index, array[index]);
